test: move tests into separate folder and add integration test example
